// ArchiCore/src/app/api-client/model/equipeResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EquipeResponse { 
    id?: number;
    codeMission?: string;
    userId?: number;
    poste?: EquipeResponse.PosteEnum;
}
export namespace EquipeResponse {
    export const PosteEnum = {
        ChefMission: 'CHEF_MISSION',
        Superviseur: 'SUPERVISEUR',
        Senior: 'SENIOR',
        Junior: 'JUNIOR',
        Assistant: 'ASSISTANT',
        Stagiaire: 'STAGIAIRE'
    } as const;
    export type PosteEnum = typeof PosteEnum[keyof typeof PosteEnum];
}