// ArchiCore/src/app/api-client/model/equipeCreateRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EquipeCreateRequest { 
    codeMission: string;
    userId: number;
    poste: EquipeCreateRequest.PosteEnum;
}
export namespace EquipeCreateRequest {
    export const PosteEnum = {
        ChefMission: 'CHEF_MISSION',
        Superviseur: 'SUPERVISEUR',
        Senior: 'SENIOR',
        Junior: 'JUNIOR',
        Assistant: 'ASSISTANT',
        Stagiaire: 'STAGIAIRE'
    } as const;
    export type PosteEnum = typeof PosteEnum[keyof typeof PosteEnum];
}